refactor: split admin_service.go into sub-domain files
Same per-sub-domain split pattern as admin_repo.go.
AdminService struct, constructor, platformStatsCache, and helper
methods (logAudit, normalizePagination) stay in core admin_service.go.

No behavior change.
